Change float to double in cloth simulation

+++ include/Simulation/Cloth.h
@@ -3,36 +3,33 @@
 #include "Models/Mesh.h"
 #include "Simulation/PointMass.h"
 #include "Simulation/Spring.h"
-#include <vector>
 #include <memory>
+#include <vector>
 
-//TODO: change float to double
 struct Cloth {
 public:
     Cloth(
-        const glm::vec3& lowerLeftCorner_,
-        const float width_,
-        const float height_,
+        const glm::dvec3& lowerLeftCorner_,
+        const double width_,
+        const double height_,
         const std::uint32_t widthPoints_,
         const std::uint32_t heightPoints_);
 
     std::shared_ptr<Mesh> mesh;
 
     void simulate(
-        float deltaTime,
+        double deltaTime,
         std::uint32_t simulationSteps,
-        std::vector<glm::vec3> accelerations
-    );
+        std::vector<glm::dvec3> accelerations);
 
     void createMassesAndSprings();
     std::shared_ptr<Mesh> createMesh();
     void recomputePositionsNormals();
 
 private:
-
-    glm::vec3 lowerLeftCorner;
-    float width;
-    float height;
+    glm::dvec3 lowerLeftCorner;
+    double width;
+    double height;
     std::uint32_t widthPoints;
     std::uint32_t heightPoints;
 
@@ -40,7 +37,7 @@ private:
     std::vector<Spring> springs;
     std::vector<glm::ivec3> triangles;
 
-    float dumping = 0.2f;
-    float density = 150.0f;
-    float ks = 5000.0f;
+    double dumping = 0.2;
+    double density = 150.0;
+    double ks = 5000.0;
 };
